# bank/src/test/resources/spec/digitalbank/openapi/DigitalBank-OpenAPI3.0.json
openapi: 3.0.0
info:
  description: Digial Bank API provides Administration and User functionality through API
    Endpoints.
  version: API TOS
  title: Digital Bank API
  termsOfService: Terms of Service
  contact:
    name: Digital Bank
    url: www.demo.io
    email: digitalbank@demo.io
  license:
    name: License of API
    url: API License URL
...
paths:
  /api/v1/account:
    get:
      tags:
        - account-controller
      summary: getAllAccounts
      operationId: getAllAccountsUsingGET
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
  /api/v1/account/checking:
    get:
      tags:
        - account-controller
      summary: getAllCheckingAccounts
      operationId: getAllCheckingAccountsUsingGET
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
  /api/v1/account/savings:
    get:
      tags:
        - account-controller
      summary: getAllSavingsAccounts
      operationId: getAllSavingsAccountsUsingGET
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
  "/api/v1/account/{id}":
    get:
      tags:
        - account-controller
      summary: getAccount
      operationId: getAccountUsingGET
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
    put:
      tags:
        - account-controller
      summary: updateAccount
      operationId: updateAccountUsingPUT
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
        - name: newName
          in: query
          description: newName
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "201":
          description: Created
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
    delete:
      tags:
        - account-controller
      summary: deleteAccount
      operationId: deleteAccountUsingDELETE
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "204":
          description: No Content
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      security:
        - Bearer_token: []
      deprecated: false
  "/api/v1/account/{id}/coowner":
    get:
      tags:
        - account-controller
      summary: getCoOwner
      operationId: getCoOwnerUsingGET
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
    put:
      tags:
        - account-controller
      summary: setCoOwner
      operationId: setCoOwnerUsingPUT
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: query
          description: userId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "201":
          description: Created
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
  "/api/v1/account/{id}/owner":
    get:
      tags:
        - account-controller
      summary: getOwner
      operationId: getOwnerUsingGET
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
  "/api/v1/account/{id}/transaction":
    get:
      tags:
        - account-controller
      summary: getAccountTransactions
      operationId: getAccountTransactionsUsingGET
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
    post:
      tags:
        - account-controller
      summary: newTransaction
      operationId: newTransactionUsingPOST
      parameters:
        - name: action
          in: query
          description: action
          required: false
          schema:
            type: string
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewTransaction"
        description: newTransaction
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "201":
          description: Created
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
  "/api/v1/account/{id}/transfer":
    post:
      tags:
        - account-controller
      summary: transferFunds
      operationId: transferFundsUsingPOST
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferFunds"
        description: transfer
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "201":
          description: Created
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
...
  /api/v1/user/account:
    get:
      tags:
        - account-controller
      summary: getAccounts
      operationId: getAccountsUsingGET
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
    post:
      tags:
        - account-controller
      summary: createAccount
      operationId: createAccountUsingPOST
      requestBody:
        $ref: "#/components/requestBodies/NewAccount"
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "201":
          description: Created
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
  /api/v1/user/account/checking:
    get:
      tags:
        - account-controller
      summary: getCheckingAccounts
      operationId: getCheckingAccountsUsingGET
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
  /api/v1/user/account/savings:
    get:
      tags:
        - account-controller
      summary: getSavingsAccounts
      operationId: getSavingsAccountsUsingGET
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
...
  "/api/v1/user/{id}/account":
    get:
      tags:
        - account-controller
      summary: getAccounts
      operationId: getAccountsUsingGET_1
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
    post:
      tags:
        - account-controller
      summary: createAccount
      operationId: createAccountUsingPOST_1
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: "#/components/requestBodies/NewAccount"
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "201":
          description: Created
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
  "/api/v1/user/{id}/account/checking":
    get:
      tags:
        - account-controller
      summary: getCheckingAccounts
      operationId: getCheckingAccountsUsingGET_1
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
  "/api/v1/user/{id}/account/savings":
    get:
      tags:
        - account-controller
      summary: getSavingsAccounts
      operationId: getSavingsAccountsUsingGET_1
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
...
servers:
  - url: //localhost:8080/bank
components:
  requestBodies:
...
    NewAccount:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NewAccount"
      description: account
      required: true
  securitySchemes:
    Bearer_token:
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    NewAccount:
      type: object
      properties:
        accountName:
          type: string
        accountTypeCode:
          type: string
        openingDeposit:
          type: number
        ownerTypeCode:
          type: string
      title: NewAccount
    NewTransaction:
      type: object
      properties:
        amount:
          type: number
        description:
          type: string
        transactionTypeCode:
          type: string
      title: NewTransaction
...
    TransferFunds:
      type: object
      properties:
        amount:
          type: number
        toAccountId:
          type: integer
          format: int64
      title: TransferFunds
...